pyproject.toml
[project]
name = "documentation-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncio>=4.0.0",
    "black>=26.1.0",
    "certifi>=2026.1.4",
    "isort>=7.0.0",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.17",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.2.1",
]
